inputFreatures: loading binary files instead of text refs#CGU-259

+++ include/caffe/ultinous/triplet_data_layer.hpp
@@ -47,7 +47,7 @@ private:
 
   typedef vector<Dtype> FeatureVector;
   vector<FeatureVector> m_inputFeatures;
-  size_t m_inputFeatureLength;
+  uint32_t m_inputFeatureLength;
 };
